+++ NEWS.md
@@ -1,3 +1,12 @@
+# lg 0.2.0.9004
+ 
+   * Added contagion test and independence tests.
+   * Added bw_simple for quick construction of bandwidth object.
+
+# lg 0.2.0.9003
+
+   * Fixed NA handling with regard to 0.2.0.9002.
+
 # lg 0.2.0.9002 
 
    * Switches to old logspline algorithm if the new fails (very rare).